全部模型及迁移文件

+++ xadmin-server/sgp/models/teacher.py
@@ -5,7 +5,7 @@ from common.core.models import DbSgpBaseModel
 
 class Teacher(DbSgpBaseModel):
     teacher_name = models.CharField(verbose_name=_("教师姓名"), max_length=255, null=True, blank=True)
-    school_id = models.IntegerField(verbose_name=_("学校"), null=True, blank=True)
+    school_id = models.BigIntegerField(verbose_name=_("学校"), null=True, blank=True)
 
     class Meta:
         verbose_name = _("Teacher")
